pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arrowhead"
version = "0.1.0"
description = "Python SDK and CLI for the Arrowhead Framework"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Arrowhead Framework", email = "info@arrowhead.eu"}
]
keywords = ["arrowhead", "iot", "industrial", "framework", "soa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: System :: Networking"
]
dependencies = [
    "requests>=2.28.0",
    "cryptography>=3.0.0",
    "click>=8.0.0",
    "flask>=2.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "rich>=12.0.0",
    "flask-cors>=4.0.0",
    "PyJWT>=2.8.0",
    "jwcrypto>=1.5.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0"
]

[project.scripts]
arrowhead = "arrowhead.cli:main"

[project.urls]
Homepage = "https://github.com/johankristianss/arrowhead"
Repository = "https://github.com/johankristianss/arrowhead"
Documentation = "https://arrowhead.eu"

[tool.setuptools.packages.find]
where = ["."]
include = ["arrowhead*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
